[aBUSE] Implement ABUSE_GET_WRITE_DATA ioctl
This enables the server to return error on write.
Previously to this commit, ABUSE_PUT_BIO on WRITE
would copy the write data into the server's
buffers but would also complete the bio with
success.

This commit implements ABUSE_GET_WRITE_DATA which
the server calls after getting a WRITE bio. The
driver then copies the write data into the
server's buffers but does not complete the
bio. The server then performs the write and
complete the bio afterwards with success or
failure with ABUSE_PUT_BIO ioctl.

Signed-off-by: Luben Tuikov <ltuikov@yahoo.com>