[BLOCK] aBUSE: compete bio with error only on put ioctl
If a bio completes with error indication from user
space, i.e. ab_result is not 0, then if the bio is
a read or write bio, complete the bio with error
and the ioctl with success. Else if the bio is
ABUSE_GET_WRITE_DATA, complete the ioctl with
success and let user space complete the bio with a
subsequent put ioctl.

Signed-off-by: Luben Tuikov <ltuikov@yahoo.com>